Reduce baby! Reduce!
